refactor: 채팅 도메인 모델과 JPA 엔티티 분리
- 순수 도메인 모델 Chat.java 추가
- JPA 엔티티 ChatEntity로 이름 변경
- 영속성 계층과 도메인 계층 분리 완료
